fix auth response

+++ app/routes/auth.py
@@ -17,18 +17,25 @@ from flask import request
 def login():
     auth = request.get_json()
     if not auth or not auth.get('email') or not auth.get('password'):
-        return make_response('oupla!', 401, {'WWW-Authenticate': 'Basic-realm= "Login required!"'})
+        return make_response('Champ du formulaire incorrect', 401,
+                             {'WWW-Authenticate': 'Basic-realm= "Login required!"'})
 
     user = User.query.filter_by(email=auth['email']).first()
     if not user:
-        return make_response('Could not verify user, Please signup!', 401,
+        return make_response('Aucun utilisateur trouvé avec cette adresse email', 401,
                              {'WWW-Authenticate': 'Basic-realm= "No user found!"'})
+
+    print({
+        'password': user.password,
+        'passwordRequest': auth.get('password')
+    })
 
     if check_password_hash(user.password, auth.get('password')):
         # Delete oldest token
         old_token = Token.query.filter_by(user_id=user.id).first()
-        db.session.delete(old_token)
-        db.session.commit()
+        if old_token:
+            db.session.delete(old_token)
+            db.session.commit()
 
         # Create and add New Token
         token = jwt.encode({'id': user.id}, app.config['SECRET_KEY'], 'HS256')
@@ -36,40 +43,13 @@ def login():
         db.session.add(token)
         db.session.commit()
 
-        return make_response(jsonify({'token': token.serialize()}), 201)
+        return make_response(jsonify({
+            'token': token.serialize(),
+            'user': user.serialize()
+        }), 201)
 
-    return make_response('Could not verify password!', 403, {'WWW-Authenticate': 'Basic-realm= "Wrong Password!"'})
+    return make_response('Mot de passe incorrect', 403, {'WWW-Authenticate': 'Basic-realm= "Wrong Password!"'})
 
-
-# @router.route("/logout/<int:user_id>", methods=["DELETE"])
-# @jwt_required()
-# def logout(user_id):
-#     user = request.get_json()
-#     token = request.get_json()
-#     if token.get('token') and user.get('user_id'):
-#         db.session.delete(token)
-#     return make_response(user_id, 204)
-
-# ACCESS_EXPIRES = timedelta(hours=1)
-#
-# jwt_redis_blocklist = redis.StrictRedis(
-#     host="localhost", port=6000, db=0, decode_responses=True
-# )
-#
-#
-# @jwt.token_in_blocklist_loader
-# def check_if_token_is_revoked(jwt_header, jwt_payload: dict):
-#     jti = jwt_payload["jti"]
-#     token_in_redis = jwt_redis_blocklist.get(jti)
-#     return token_in_redis is not None
-
-
-# @router.route("/logout", methods=["DELETE"])
-# @jwt_required()
-# def logout():
-#     jti = get_jwt()["jti"]
-#     # jwt_redis_blocklist.set(jti, "", ex=ACCESS_EXPIRES)
-#     return jsonify(msg="Access token revoked")
 
 @router.route('/logout', methods=["GET"])
 def logout():
@@ -98,7 +78,7 @@ def forgot_password():
     msg.body = url
     mail.send(msg)
 
-    return make_response('Email sent at ' + email, 200)
+    return make_response('Email envoyé à ' + email, 200)
 
 
 @router.route('/reset-password/<string:token>', methods=["POST"])
